feat(CC-0059): Add security and correctness linters to golangci-lint (#176)
- Enable six new linters in .golangci.yml: bodyclose, errorlint,
exhaustive, gosec, nilerr, and noctx for security and correctness
analysis across all Go modules
- Suppress gosec G101 (hardcoded credential heuristics) in test files
to avoid false positives on fixture connection strings and tokens
- Add targeted nolint directives for justified exceptions: envtest
binary path check (G703), credential rotation script const name
(G101), and test utility TLS dial without context (noctx)
- Change SimulateMariaDBReady replicas parameter from int to int32
to eliminate an unsafe integer conversion flagged by gosec
- Document all twelve enabled linters in docs/reference/ci-workflow.md
with category, description, and originating feature ID

AI-assisted: Claude Code
On-behalf-of: @SAP christian.berendt@sap.com

+++ internal/common/testutil/simulators/simulators.go
@@ -77,13 +77,13 @@ func setUnstructuredReadyStatus(
 // SimulateMariaDBReady updates a MariaDB resource's status to indicate
 // readiness by setting the Ready condition to True, replicas, and
 // currentPrimaryPodIndex.
-func SimulateMariaDBReady(ctx context.Context, c client.Client, key client.ObjectKey, replicas int) error {
+func SimulateMariaDBReady(ctx context.Context, c client.Client, key client.ObjectKey, replicas int32) error {
 	mariadb := &mariadbv1alpha1.MariaDB{}
 	if err := c.Get(ctx, key, mariadb); err != nil {
 		return fmt.Errorf("getting MariaDB %s: %w", key, err)
 	}
 
-	mariadb.Status.Replicas = int32(replicas)
+	mariadb.Status.Replicas = replicas
 	primaryIdx := 0
 	mariadb.Status.CurrentPrimaryPodIndex = &primaryIdx
